Desktop/IDInfoExtraction-main/requirements.txt
pytesseract~=0.3.13
pillow~=11.0.0
Flask~=3.1.0
googletrans~=3.0.0